Move logic to plugin module
