+++ setup.cfg
@@ -37,7 +37,7 @@ test =
 
 [options.entry_points]
 pytest11 =
-    dbt=pytest_dbt_core.fixtures
+    dbt=pytest_dbt_core.plugin
 
 [flake8]
 ignore = E226,E302,E41,W504,W503
